move mocked_s3_sink_client to global conftest
