+++ tests/datenkompass/conftest.py
@@ -1,5 +1,5 @@
-from typing import TYPE_CHECKING, cast
-from unittest.mock import MagicMock, patch
+from typing import cast
+from unittest.mock import MagicMock
 
 import pytest
 from pytest import MonkeyPatch
@@ -33,9 +33,6 @@ from mex.extractors.datenkompass.models.item import DatenkompassActivity
 from mex.extractors.datenkompass.models.mapping import DatenkompassMapping
 from mex.extractors.settings import Settings
 from mex.extractors.utils import load_yaml
-
-if TYPE_CHECKING:
-    from collections.abc import Generator
 
 
 @pytest.fixture
@@ -448,15 +445,3 @@ def mocked_provider(monkeypatch: MonkeyPatch) -> MagicMock:
     monkeypatch.setattr(provider, "fetch", fake_provider.fetch)
 
     return fake_provider
-
-
-@pytest.fixture  # needed for hardcoded upload to S3.
-def mocked_boto() -> Generator[MagicMock]:
-    """Mock a S3 session client to write the jsons to."""
-    with patch("boto3.Session") as mock_session_class:
-        mock_s3_client = MagicMock()
-        mock_session_instance = MagicMock()
-        mock_session_instance.client.return_value = mock_s3_client
-        mock_session_class.return_value = mock_session_instance
-
-        yield mock_s3_client
