+++ tests/datenkompass/test_main.py
@@ -4,7 +4,7 @@ from mex.extractors.pipeline import run_job_in_process
 
 
 @pytest.mark.usefixtures(
-    "mocked_backend_datenkompass", "mocked_provider", "mocked_boto"
+    "mocked_backend_datenkompass", "mocked_provider", "mocked_s3sink_client"
 )
 def test_run() -> None:
     assert run_job_in_process("datenkompass")
